+++ src/main/java/com/prgrms/nabmart/global/auth/controller/AuthControllerAdvice.java
@@ -1,5 +1,8 @@
-package com.prgrms.nabmart.global.auth.exception;
+package com.prgrms.nabmart.global.auth.controller;
 
+import com.prgrms.nabmart.global.auth.exception.AuthException;
+import com.prgrms.nabmart.global.auth.exception.OAuthUnlinkFailureException;
+import com.prgrms.nabmart.global.auth.exception.UnAuthenticationException;
 import com.prgrms.nabmart.global.util.ErrorTemplate;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
